requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
Flask==2.3.2
flatbuffers==23.5.26
frozenlist==1.3.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
numpy==1.24.4
onnxruntime==1.15.1
openai==0.27.8
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==4.23.3
python-dotenv==1.0.0
requests==2.31.0
sympy==1.12
tqdm==4.65.0
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
googletrans==4.0.0-rc1
moviepy==1.0.3
librosa==0.10.0.post2
openai-whisper==20230314